r.sample.category: create points per category in raster map
Basic implementation works and is tested (Bash test included) but there is a lot of to improve (see TODOs in code and documentation).

Documentation contains example and picture but description and use cases are missing.

Code goes through 'pep8 file.py' and 'pylint -r n file.py' in an acceptable way.


git-svn-id: https://svn.osgeo.org/grass/grass-addons@65357 15284696-431f-4ddb-bdfa-cd5b030d7da7