v.lidar.mcc: copy instead of merging into an empty vector
When vector map is created with v.edit it is empty and 2D. Merging 3D
vector map into the empty one using v.patch produces 2D vector map
and a warning.

Now both ground and also non-ground vetor maps are 3D.
The price is one if in the main code - the inner loop (negligible here).


git-svn-id: https://svn.osgeo.org/grass/grass-addons@66242 15284696-431f-4ddb-bdfa-cd5b030d7da7