r.forestfrag: simplify pff (pairs) expressions
Expression is generated by a function. Indexing more straightforward.
Using binary (here same as logical) operators instead of
plus and if (or), and multiplication (and).
Using mapcalc function interface which passes everything
on stdin to avoid need for a tmp file.

There is a very small performance gain for large window sizes.


git-svn-id: https://svn.osgeo.org/grass/grass-addons@68750 15284696-431f-4ddb-bdfa-cd5b030d7da7