r.in.usgs: reuse extracted .img files and imported rasters
If extracted files or raster maps are present (currently preserved only with -k flag),
use them in the import process.

This also improved and fixes several error and info messages, e.g. for cases
when there is no space left on the device and reports the state when no tiles are present
for patching.