r.in.usgs: Do not leave processes behind
In case there are tiles skipped at the end of the list (while continue),
we need to check for running processes and join if we are at the end of
the loop, otherwise we leave processes behind which may not be finished
in time for further use and will try to communicate with the manager
when it does not exist anymore.

Also adding early check and a message with suggested next step
for case when the directory for files does not exist.