Add CI with build and test (#205)
This adds a new set of scripts for build, i.e., it does not use the current build script because they mix 6 and 7 and are not parametrized enough.

The build also does not return non-zero in case of failure. Specifically, the make step does not return non-zero in case of failure because module failure is not counted as overall failure. (This does not go well with GitHub Actions or Travis style CI).

To avoid problems with getting the right Python package version combo, version is not required in the requirements.txt file. (This may need to be addressed in the future.)

There is only a small number of tests, i.e., the test coverage is small, but they are running now. However, most of the tests are failing. (Many can be probably fixed by reviewing the datasets used.)