g.download.location: Make unzip, untar more robust (#301)
The module can now deal with location being only one of the nested directories for cases
when archiving program creates one extra. It recurses one level into the structure
(can be modified in the code) which adds additional level to what the Python libraries are doing
(which actually makes the code more complex for us).

xz is among the recognized extensions.

Several special cases and error are now handled gracefully with error being reported
for single file in an archive (typically another archive, but leaving that aside)
and for file being incomplete (likely not fully downloaded).

Functions raise an unified exception, main code now calls fatal when that happens.

Filenames without extension don't cause traceback, but cause an error to be reported.