CI: Add extra addon-specific dependencies
For addons it is common to use scipy and R (and possibly other things).
This adds these to GitHub Action workflow so that the test of modules which use these
dependencies can run. It is also an intersting list of runtime dependencies which might be optional
for core and required for some addons.

These extra dependencies are installed only after core and addons compile which should
test for the compilating not requiring these extra dependencies.