CI: Add extra addon-specific dependencies (#468)
For addons it is common to use pandas, scipy (and others) and R (Rscript).
This adds these to GitHub Action workflow so that the test of modules which use these
dependencies can run. It is also an intersting list of runtime dependencies which might be optional
for core and required for some addons.

These extra dependencies are installed only after core and addons compile which should
test for the compilation not requiring these extra dependencies.

Using R v4 (latest) to make packages work and because that's probably the best
practice (and simplest) for general R dependency management.

Notably, GDAL Python binding are missing (e.g., v.import, WMS in core, v.in.pygbif in addons)
because better dependency handling is needed there. 

Increase min required test success percentage to 60% (from 30%).