g.download.location: Add path option for compatibility with core module
This adds path option for compatibility with the version of the module in v8.
It is an additional option, so the original dbase still works.
Same code can now be used for v7 and v8, but the original code for v7 still works
(v8 code should be updated to use the version in core, not the addon).