Update name of core branch from master to main
The development/default branch in the core repo will be renamed from master to main.
This includes updates of the CI, comments and documentation.
It removes the mention of the master/main branch when not necessary.

Notably, this does not update the code for OSGeo cron jobs because name of branch is used there additionally as a directory name and it is part of some non-scripted part of the setup.
This needs to be changed together with the setup on the OSGeo server.