r.in.usgs: Use system user cache dir
Instead of a required output/work/cache dir, use a subdirectory of the user cache dir according to the standards of each operating system.

Linux/Freedesktop uses .cache and we use grass as the application name, then version string (from g.version), and then name of the module.
MacOS is the same, but in Library/Caches. For Windows, we use APPDATA as in the startup script with Cache as an additional subdirectory.