r.vif: Add r.random seed handling (#712)
* Update r.vif.py

Added -s flag and seed option. This follows the practice for all modules that generate or use random data. For this addon that is when the user opts to use randomly sampled raster values to compute the vif. Note that this breaks the previous use of the -s flag, which was used to set the option to only print selected variables to screen. Now the -v flag is used to set that option.