r.boxplot better error handling
The r.colors.out, used to extract the color table of the zonal map, sometimes returns a color table with floating category id's instead of integers. I assume this is when the raster layer has no color table defined? The change checks if the category id's are of the type integer. If not, a more informative error message is returned.
In addition, a new option to set the boxplot width(s) is added, and the manual page is updated accordingly.