Update r.boxplot.py
Add option to plot the median of the whole raster as a line, and the IQR of the whole raster as a band (incl. possibility to define their colors)
The function will set the region to match that of the zonal raster. Optionally, the user can tell the function to respect the region settings.
Rearrangement of tabs and other minor fixes.