Update r.bioclim.html (#764): check the unit of measure of the precipitation rasters
I added a couple of lines highliting that the bio18 and the bio19 are calculated rounding the precipitation values to the closest integer. I think is helpful to suggest to check the unit of measure of the precipitation rasters for people that are not used to work with climatic data. I had this problem with the CHELSA database that provides pricipitation in kg m-2 s-1, with values very close to 0 that lead to a raster of 0s for bio18 and bio19.