r.series.boxplot: new addon to create boxplots of series of rasters (#779)
New addon that create boxplots of a series of input raster layers. The boxplot(s) show temporal patterns across a region, and for each time step the variation within the region. There are options to format the boxplot(s), and the plot can be saved as png, svg or other matplotlib supported format.