r.bioclim: End with success when precipitation is missing
The precipitation parameter is optional, so the process return code should be 0, not 1, as no error occured.

A return statement is more appropriate here than a direct exit call because this is a main function. Using return is consistent with the other branch which ends with the end of the function (an implicit return).