r.bioclim: End with success when precipitation is missing (#886)
The precipitation parameter is optional, so the process return code should be 0, not 1, as no error occurred.

A return statement is more appropriate here than a direct exit call because this is a main function. Using return is consistent with the other branch which ends with the end of the function (an implicit return).

Remove sys import which would be unused.