r.niche.similarity: address #897 (#898)
* Partially address #897

This small PR solves the Attribute error part reported in #897

* address the type error in printing

This solves the second error. The text file does not look great though...

* Update src/raster/r.niche.similarity/r.niche.similarity.py

Co-authored-by: Anna Petrasova <kratochanna@gmail.com>

* update r.niche.similarity bugfix

Changes to solve the errors reported in #897. The printing in the csv file is fixed as well (was caused by earlier solution, reverted to old code used to print to csv).

---------

Co-authored-by: Anna Petrasova <kratochanna@gmail.com>
Co-authored-by: Paulo van Breugel <paulo@ecodiv.earth>