r.fuzzy.set: correction of sshape formula in fuzzy.c (#931)
This PR corrects two issues with how the S-shape rescaling was implemented for negative values of the shape parameter. This includes the issue reported in #931.

In addition, in the manual page r.fuzzy.set.html, the description of the equations were updated to ensure that they match the equations as implemented in the code. And an issue with a wrongly formatted URL was corrected.