Replaced num of params for setitem method (#993)
In file: road_tables.py, class: RoadTable, there is a special method __setitem__ that contains an unexpected number of parameters. Each call to this method will result in a TypeError. iCR suggested that special methods should have an expected number of parameters. More infomation can be found in the Python documentation on the various special method.

Here, the error is being thrown, because the __setitem__ method (the one under the hood), is expecting another argument called value- because we defined our __setitem__ method as such. Which is why, we must follow documentation and implement the methods accordingly. In this case, we can simply extract the index and name from the parameter which can be a tuple.

Signed-off-by: fazledyn-or <ataf@openrefactory.com>