Added check for square raster cells
Maxent only accepts raster cells with equal ns and ew resolution. Added code that checks the region setting, and gives a fatal message if ns resolution is not ew resolution. Optionally, the user can set the -e flag. This will automatically adjust the resolution of the largest side to match that of the smallest side.