when exporting swd file, also export a proj. file with the crs info.
I plan to use this proj. file in r.maxent.train to check if the crs of the data used for the training matches that of the current mapset. If not, this proj. file will make it possible to import the output of r.maxent.train with reprojection (with warning). I can still have r.maxent.train issue a warning if there is no proj file, and let the user explicitly use the -o flag to override the projection check.