r.maxent.train: a new addon to train a model using Maxent (#1012)
* r.maxent.train: a new addon to train a model using Maxent

The addon, essentially a wrapper to the Maxent software, can be used for the training of a Maxent model. It complements the addons v.maxent.swd, r.out.maxent_swd and r.maxent.lambda. Together they allow one to carry out a good part of the model workflow, from data preparation to training to prediction. An additional addon to use the model outcome for predictions using alternative prediction layers, also using Maxent in the background, is in development.

Co-authored-by: Stefan Blumentrath <stefan.blumentrath@gmx.de>
Co-authored-by: Edouard Choinière <27212526+echoix@users.noreply.github.com>
Co-authored-by: Veronica Andreo <veroandreo@gmail.com>