r.in.ahn: new addon to import the AHN DTM or DSM (#1043)
Addon to import the 'Actueel Hoogtebestand Nederland' in GRASS GIS. The AHN consist of high resolution (0.5m) DTM and DSM layers for the Netherlands. The addon uses r.in.wcs in the background. It sets the extent and resolution so that the imported data aligns with the original AHN data, making it, in that respect, like importing the raster layer with r.in.gdal. The user can opt to set the region back to the original region settings after the layer is imported. 



Co-authored-by: Anna Petrasova <kratochanna@gmail.com>