t.rast.line addon
New addon to draw trendlines based on the raster layers from a spatio-temporal raster database, showing for each time step the average values of the raster layer. If a zonal map is provided, trend lines are drawn showing for each time step the average value of each of the categories of the zonal raster layer.