Better error handling and added illustration
Illustration of an use case; comparing seasonal patterns in FCover between land cover types.

Use try / exception so when strds does not exists, the module fails immediately, avoiding all the extra lines in error message. This is not an ideal solutions as it now gives two error message, but not sure how to do this otherwise.

Added clean-up function to clean up intermediate layers.