r.in.vect: new addon to import a vector map as raster layer (#1143)
An new addon offering an alternative to importing a vector layer and then converting the vector to raster. With the addon, the input vector file is converted to a raster layer using gdal.Rasterize. The resulting raster is imported in GRASS GIS. This can speed up the whole process considerably if one only wants to work with the raster.

If the CRS of the vector layer differs from the current mapset, it will first be reprojected using ogr2ogr. By default, the addon respects the region settings, but the user can opt to create a raster layer matching the extent of the vector map.



---------

Co-authored-by: Veronica Andreo <veroandreo@gmail.com>