r.maxent.predict: add option to provide folder with input raster layers
Add option to provide folder with input raster layers - time saver if the same layers are required multiple times.
Improvement check if right layers/variables are used as input.
Added check if provided layer names exist in more than one mapset (only if mapset is not provided). This will result in error, except if the layers are also in the current mapset (because in that case GRASS chooses those anyway).