r.maxent.predict: change parameter names to avoid conflicts when scripting (#1238)
* bugfixes and improved error handling

* Add / improve error messages

Add progress indicator
Improve error messages

* r.maxent.predict: add option to provide folder with input raster layers

Add option to provide folder with input raster layers - time saver if the same layers are required multiple times.
Improvement check if right layers/variables are used as input.
Added check if provided layer names exist in more than one mapset (only if mapset is not provided). This will result in error, except if the layers are also in the current mapset (because in that case GRASS chooses those anyway).

* Rename parameters to avoid conflicts