v.boxplot and r.boxplot: solve issue with running addons in VS Code
Running the addons in VS Code results in an error related with `matplotlib.use("WXAgg")`. A possible solution is using another backend like `WebAgg`. However, running VS Code probably means one wants to print the resulting graph to file. As in that case, the backend is not needed, we can just avoid loading it when the option to save the graph to file is selected.